#413 re-separate zocalo_interaction and callback
