+++ src/artemis/external_interaction/callbacks/fgs/fgs_zocalo_callback.py
@@ -1,26 +1,17 @@
-import getpass
 import math
-import queue
-import socket
 import time
 from typing import Callable
 
-import workflows.recipe
-import workflows.transport
-import zocalo.configuration
 from bluesky.callbacks import CallbackBase
-from workflows.transport import lookup
 
-import artemis.log
 from artemis.external_interaction.callbacks.fgs.fgs_ispyb_callback import (
     FGSISPyBHandlerCallback,
 )
 from artemis.external_interaction.exceptions import ISPyBDepositionNotMade
+from artemis.external_interaction.zocalo_interaction import ZocaloInteractor
 from artemis.log import LOGGER
 from artemis.parameters import ISPYB_PLAN_NAME, FullParameters
 from artemis.utils import Point3D
-
-TIMEOUT = 90
 
 
 class FGSZocaloCallback(CallbackBase):
@@ -45,12 +36,12 @@ class FGSZocaloCallback(CallbackBase):
         self.grid_position_to_motor_position: Callable[
             [Point3D], Point3D
         ] = parameters.grid_scan_params.grid_position_to_motor_position
-        self.zocalo_env = parameters.zocalo_environment
         self.processing_start_time = 0.0
         self.processing_time = 0.0
         self.results = None
         self.xray_centre_motor_position = None
         self.ispyb = ispyb_handler
+        self.zocalo_interactor = ZocaloInteractor(parameters.zocalo_environment)
 
     def event(self, doc: dict):
         LOGGER.debug(f"\n\nZocalo handler received event document:\n\n {doc}\n")
@@ -61,7 +52,7 @@ class FGSZocaloCallback(CallbackBase):
             if self.ispyb.ispyb_ids[0] is not None:
                 datacollection_ids = self.ispyb.ispyb_ids[0]
                 for id in datacollection_ids:
-                    self._run_start(id)
+                    self.zocalo_interactor.run_start(id)
             else:
                 raise ISPyBDepositionNotMade("ISPyB deposition was not initialised!")
 
@@ -71,12 +62,12 @@ class FGSZocaloCallback(CallbackBase):
             raise ISPyBDepositionNotMade("ISPyB deposition was not initialised!")
         datacollection_ids = self.ispyb.ispyb_ids[0]
         for id in datacollection_ids:
-            self._run_end(id)
+            self.zocalo_interactor.run_end(id)
         self.processing_start_time = time.time()
 
     def wait_for_results(self, fallback_xyz: Point3D):
         datacollection_group_id = self.ispyb.ispyb_ids[2]
-        raw_results = self._wait_for_result(datacollection_group_id)
+        raw_results = self.zocalo_interactor.wait_for_result(datacollection_group_id)
         self.processing_time = time.time() - self.processing_start_time
         # _wait_for_result returns the centre of the grid box, but we want the corner
         self.results = Point3D(
@@ -97,103 +88,3 @@ class FGSZocaloCallback(CallbackBase):
 
         LOGGER.info(f"Results recieved from zocalo: {self.xray_centre_motor_position}")
         LOGGER.info(f"Zocalo processing took {self.processing_time}s")
-
-    def _get_zocalo_connection(self, env: str = "artemis"):
-        zc = zocalo.configuration.from_file()
-        zc.activate_environment(env)
-
-        transport = lookup("PikaTransport")()
-        transport.connect()
-
-        return transport
-
-    def _send_to_zocalo(self, parameters: dict):
-        transport = self._get_zocalo_connection(self.zocalo_env)
-
-        try:
-            message = {
-                "recipes": ["mimas"],
-                "parameters": parameters,
-            }
-            header = {
-                "zocalo.go.user": getpass.getuser(),
-                "zocalo.go.host": socket.gethostname(),
-            }
-            transport.send("processing_recipe", message, headers=header)
-        finally:
-            transport.disconnect()
-
-    def _run_start(self, data_collection_id: int):
-        """Tells the data analysis pipeline we have started a grid scan.
-        Assumes that appropriate data has already been put into ISPyB
-
-        Args:
-            data_collection_id (int): The ID of the data collection representing the
-                                    gridscan in ISPyB
-        """
-        artemis.log.LOGGER.info(
-            f"Submitting to zocalo with ispyb id {data_collection_id}"
-        )
-        self._send_to_zocalo({"event": "start", "ispyb_dcid": data_collection_id})
-
-    def _run_end(self, data_collection_id: int):
-        """Tells the data analysis pipeline we have finished a grid scan.
-        Assumes that appropriate data has already been put into ISPyB
-
-        Args:
-            data_collection_id (int): The ID of the data collection representing the
-                                    gridscan in ISPyB
-
-        """
-        self._send_to_zocalo(
-            {
-                "event": "end",
-                "ispyb_wait_for_runstatus": "1",
-                "ispyb_dcid": data_collection_id,
-            }
-        )
-
-    def _wait_for_result(
-        self, data_collection_group_id: int, timeout: int = TIMEOUT
-    ) -> Point3D:
-        """Block until a result is received from Zocalo.
-        Args:
-            data_collection_group_id (int): The ID of the data collection group representing
-                                            the gridscan in ISPyB
-
-            timeout (float): The time in seconds to wait for the result to be received.
-        Returns:
-            Returns the centre of the grid box with the strongest diffraction, i.e.,
-            which contains the centre of the crystal and which we want to move to.
-        """
-        transport = self._get_zocalo_connection(self.zocalo_env)
-        result_received: queue.Queue = queue.Queue()
-
-        def receive_result(
-            rw: workflows.recipe.RecipeWrapper, header: dict, message: dict
-        ) -> None:
-            artemis.log.LOGGER.info(f"Received {message}")
-            recipe_parameters = rw.recipe_step["parameters"]
-            artemis.log.LOGGER.info(f"Recipe step parameters: {recipe_parameters}")
-            transport.ack(header)
-            received_group_id = recipe_parameters["dcgid"]
-            if received_group_id == str(data_collection_group_id):
-                result_received.put(Point3D(*reversed(message[0]["centre_of_mass"])))
-            else:
-                artemis.log.LOGGER.warn(
-                    f"Warning: results for {received_group_id} received but expected \
-                        {data_collection_group_id}"
-                )
-
-        workflows.recipe.wrap_subscribe(
-            transport,
-            "xrc.i03",
-            receive_result,
-            acknowledgement=True,
-            allow_non_recipe_messages=False,
-        )
-
-        try:
-            return result_received.get(timeout=timeout)
-        finally:
-            transport.disconnect()
